Eric develop rest

See merge request s0561491/vocabduel!58

+++ Game.impl/src/main/java/de/htwberlin/Game/impl/GameDao.java
@@ -16,5 +16,5 @@ public interface GameDao {
     public void deleteGame(Game game);
 
     @Transactional
-    public Game getGameById(Long gameId);
+    public Game getGameById(int gameId);
 }
